requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asyncer==0.0.8
attrs==25.3.0
autogen-agentchat @ git+https://github.com/patrickstolc/autogen.git@80da18f4500c043b206387a660ba6ac95b906c85
beautifulsoup4==4.13.4
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
eval_type_backport==0.2.2
exceptiongroup==1.3.0
fix-busted-json==0.0.18
FLAML==2.3.5
h11==0.16.0
httpcore==1.0.9
httpx==0.27.2
idna==3.10
jiter==0.10.0
jsonpath-python==1.0.6
lxml==5.4.0
mistralai==1.2.3
mypy_extensions==1.1.0
numpy==1.26.4
ollama==0.3.3
openai==1.82.0
outcome==1.3.0.post0
packaging==25.0
pyautogen==0.9.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.1.0
pywin32==310
regex==2024.11.6
requests==2.32.3
selenium==4.33.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
termcolor==3.1.0
tiktoken==0.9.0
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
websocket-client==1.8.0
wsproto==1.2.0
